+++ deeplearning/src/main/java/fr/cyu/smartread/deeplearning/activations/Relu.java
@@ -1,4 +1,4 @@
-package fr.cyu.smartread.deeplearning.activation;
+package fr.cyu.smartread.deeplearning.activations;
 
 import org.ejml.data.DMatrixRMaj;
