+++ deeplearning/src/test/java/fr/cyu/smartread/deeplearning/activations/AbstractActivationTest.java
@@ -1,4 +1,4 @@
-package fr.cyu.smartread.deeplearning.activation;
+package fr.cyu.smartread.deeplearning.activations;
 
 import org.ejml.EjmlUnitTests;
 import org.ejml.data.DMatrixRMaj;
